+++ scripts/speakers.sh
@@ -2,8 +2,8 @@
 
 set -e
 
-amixer -c 0 set 'Headphone' 2%
+amixer -c 0 set 'Headphone' 2% off
 
-amixer -c 0 set 'Line Out' 80%
+amixer -c 0 set 'Line Out' 80% on
 
 echo "Speakers on"